+++ packages/schemas/src/lib/1-a1/1-0-0.ts
@@ -1,11 +1,8 @@
 import { JSONSchema7 } from 'json-schema';
 
+import { schema as base } from '../0-base/1-0-0';
 import { resolveGitHubPath } from '../helpers';
-import {
-  schema as base,
-  credentialSubjectWithAssertions,
-  clause,
-} from '../0-base/1-0-0';
+import { credentialSubject, assertion } from '../schema-helpers';
 
 export const schema: JSONSchema7 = {
   ...base,
@@ -19,28 +16,21 @@ export const schema: JSONSchema7 = {
   properties: {
     ...base.properties,
 
-    credentialSubject: credentialSubjectWithAssertions({
-      type: 'object',
-      required: ['and'],
-      properties: {
-        and: {
-          type: 'array',
-          minItems: 2,
-          maxItems: 2,
-          items: [
-            clause({
-              src: '$has_web2_account',
-              op: '==',
-              dst: 'true',
-            }),
-            clause({
-              src: '$has_web3_account',
-              op: '==',
-              dst: 'true',
-            }),
-          ],
-        },
-      },
+    credentialSubject: credentialSubject({
+      assertions: assertion.and({
+        items: [
+          assertion.clause({
+            src: '$has_web2_account',
+            op: '==',
+            dst: 'true',
+          }),
+          assertion.clause({
+            src: '$has_web3_account',
+            op: '==',
+            dst: 'true',
+          }),
+        ],
+      }),
     }),
   },
 };
