+++ packages/schemas/src/lib/2-a2/1-0-0.ts
@@ -1,11 +1,8 @@
 import { JSONSchema7 } from 'json-schema';
 
+import { schema as base } from '../0-base/1-0-0';
 import { resolveGitHubPath } from '../helpers';
-import {
-  schema as base,
-  credentialSubjectWithAssertions,
-  clause,
-} from '../0-base/1-0-0';
+import { credentialSubject, assertion } from '../schema-helpers';
 
 export const schema: JSONSchema7 = {
   ...base,
@@ -19,33 +16,26 @@ export const schema: JSONSchema7 = {
   properties: {
     ...base.properties,
 
-    credentialSubject: credentialSubjectWithAssertions({
-      type: 'object',
-      required: ['and'],
-      properties: {
-        and: {
-          type: 'array',
-          minItems: 3,
-          maxItems: 3,
-          items: [
-            clause({
-              src: '$verified_discord_account',
-              op: '>',
-              dst: '0',
-            }),
-            clause({
-              src: '$has_joined',
-              op: '==',
-              dst: 'true',
-            }),
-            clause({
-              src: '$discord_guild_id',
-              op: '==',
-              dst: 'abc',
-            }),
-          ],
-        },
-      },
+    credentialSubject: credentialSubject({
+      assertions: assertion.and({
+        items: [
+          assertion.clause({
+            src: '$verified_discord_account',
+            op: '>',
+            dst: '0',
+          }),
+          assertion.clause({
+            src: '$has_joined',
+            op: '==',
+            dst: 'true',
+          }),
+          assertion.clause({
+            src: '$discord_guild_id',
+            op: '==',
+            dst: 'abc',
+          }),
+        ],
+      }),
     }),
   },
 };
