+++ packages/schemas/src/lib/schema-helpers/index.ts
@@ -1,0 +1,2 @@
+export { credentialSubject } from './credential-subject';
+export * as assertion from './assertion';
